+++ src/app/api/portal/refreshments/route.ts
@@ -12,7 +12,6 @@ import type {
   StockStatus,
   RefreshmentType,
   RefreshmentCategory,
-  UnitType,
 } from "@prisma/client";
 
 // ----------------------------------------------------------------------
