+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import * as React from "react";
 import type { Metadata } from "next";
 import Navbar from "@/components/Navbar";
 import Footer from "@/components/Footer";
-import { IBM_Plex_Sans_Thai } from "next/font/google";
 import getCurrentUser from "@/actions/userActions";
+import { IBM_Plex_Sans_Thai } from "next/font/google";
 
 import "./globals.css";
 import Providers from "./providers";
